backend/requirements.txt
Django
django-cors-headers
djangorestframework
mysqlclient
pillow